requirements.txt
flask
flask-cors
ultralytics
torch
pillow
numpy